decouple ValueObjectRemoverDocBlockRector
